Adjust the path to the shoot CA certificate in the blackbox exporter (#5947)
The blackbox exporter, which is running as a sidecar of the prometheus
in the shoot control plane namespace is used to probe the shoot api server
via its external endpoint. See the red path in the documentation:
https://github.com/gardener/gardener/blob/dc1abd68d3e04679cf810d86086611cbbefa0a66/docs/monitoring/connectivity.md

The TLS connection to the api server does not work which results in
false positive alerts: "ApiServerNotReachable".

The underlying error message is:

```
unable to load specified CA cert /var/run/secrets/shoot-ca/ca.crt:
  open /var/run/secrets/shoot-ca/ca.crt: no such file or directory
```

The blackbox exporter does not have access to the CA certificate that
it needs to verify the certificate of the shoot API server, and hence
the TLS connection fails.

This issue is related to the CA rotation epic, which introduced a different
name for the certificate to support the rotation process.

The path is pointing to this volume:
https://github.com/gardener/gardener/blob/dc1abd68d3e04679cf810d86086611cbbefa0a66/charts/seed-monitoring/charts/core/charts/prometheus/templates/prometheus.yaml#L196
where the file name changed from "ca.crt" to "bundle.crt".

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>