`kubeapiserver.Wait` function waits for all pods to be terminated
Earlier, the `Wait` function was already returning in case all replicas were updated. However, in this case there might still have been `Pod`s in `Terminating` state. The intention of this function is to make sure the deployment is fully updated, meaning that there are only most recent `Pod`s in the system.