Deflake managedresource controller integration test (#5953)
* Rename: files and test package

* Rework: managedresource integration test setup

- create dedicated namespace per test suite run (allows running suite
  in parallel on existing cluster)
- use DeferCleanup which is much easier to follow
- drop unneeded JustBeforeEach blocks
- use same schemes as in real life
- simplify secret creation

* Simplify: drop compareResource in favor of DeepEqual

* Enhance: use real matchers with Eventually instead of bool

- makes output for test failures much nicer and easier to understand
- makes condition testing more precise by checking the reason as well

* Fix: test case for DeleteOnInvalidUpdate

It was rather testing the cleanup of old resources and the DeleteOnInvalidUpdate mode
because the controller was never able to apply the new resource because of missing
TypeMeta.

* Rework: creation of test ManagedResource

- use unique resource names specific to each test case to prevent
  cascading failures because of AlreadyExistsErrors
- create ManagedResource and Secret in top-level BeforeEach
- deduplicates a lot of setup and teardown code and harmonizes
  assertions

* Deflake: should set ManagedResource to unhealthy

There was a missing To(Succeed()) when removing the test finalizer. Because
of a stale in-memory object it might have happened, that the test tried to
patch including the controller's finalizer which had already been removed
from the actual ManagedResource. Adding a finalizer to an object which is
already in deletion is denied by API server. The missing To(Succeed())
caused that such a failure would go unnoticed and the ManagedResource would
not be deleted in the end, because the test finalizer was still present.

Now, the test finalizer is removed with optimistic locking and retries to
make sure we can successfully remove it and the ManagedResource is gone in
the end.

* Deflake: test failures caused by conflict errors

Stop using optimistic locking for updating test resources and switch to
Patch wherever possible. Where not possible add retries.

* Deflake: reduce requeue time on pending deletion

* Change back to Consistently