Adapt generic `ControlPlane` actuator
Earlier, it was using `[]admissionregistrationv1.MutatingWebhook`, however this was cumbersome and required creating the `MutatingWebhookConfiguration` multiple times.

Additionally, now that we have a separate reconciler for managing the CA bundle in webhooks, it is more convenient to always use `MutatingWebhookConfiguration`. Particularly, the `AtomicShootWebhookConfig` passed to the generic actuator is an `atomic.Value` which is shared with the CA bundler reconciler so that we can make sure that the webhook config for newly created shoots always contains the most up-to-date CA bundle right away.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>