[release-v1.45] Reverse checks for invalidating clients (#5997)
* Reverse checks for invalidating clients

Earlier, the client was never invalidated/refreshed if the current token
was expired or invalidated because the `DiscoverVersion()` call never
succeeded (`Unauthorized`).
Now the hash of the client is checked first to ensure that a new client
is always created when the hash changed.

Also, earlier the client was not refreshed when the server version
changed (just a log statement was printed). This is now done as well.

* Do not refresh clients on version change

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>