[GEP-18] Use new secrets manager for extension CA and server secret generation, auto-rotate those every 30d (#6003)
* Introduce `secretsmanager.IgnoreOldSecretsAfter` option

The logic for when old secrets should be dropped after rotation from the
seed controller is extracted into a dedicated secrets manager option.
This will be reused in the extensions for dropping old CAs after
auto-rotation similar to the seed CA.

* Drop unused `webhook.Webhook.Kind` field

This `Kind` field is not used in the `extensions/pkg/webhook` package but only in `extensions/pkg/webhook/controlplane`.

* Split up RegisterWebhooks into two functions

Previously, RegisterWebhooks used to build the webhook configs
and reconcile the seed webhook config as well. However, the shoot
webhook config would need to be reconciled separately. This is
confusing and needs to be split for the purpose of this PR.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Refactor webhook certificate generation

In the future, there will be a reconciler which manages the certificate, so we need to prepare accordingly.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Add reconciler for certificates

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Add reloader for certificates

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Add the new reconcilers as `Runnable`s to manager

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Integration test for certificate reconciler and reloader

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Adapt generic `ControlPlane` actuator

Earlier, it was using `[]admissionregistrationv1.MutatingWebhook`, however this was cumbersome and required creating the `MutatingWebhookConfiguration` multiple times.

Additionally, now that we have a separate reconciler for managing the CA bundle in webhooks, it is more convenient to always use `MutatingWebhookConfiguration`. Particularly, the `AtomicShootWebhookConfig` passed to the generic actuator is an `atomic.Value` which is shared with the CA bundler reconciler so that we can make sure that the webhook config for newly created shoots always contains the most up-to-date CA bundle right away.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Drop `shootWebhookReconciler` `Runnable` from `provider-local`

This is not needed anymore now that we have dedicated reconcilers for manging the certificates and webhook configurations.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Wait for webhook reconciler to populate CA bundle before triggering webhooks

Now that the webhook CAs are populated by a reconciler and are not part of the webhook config at their initial creation, we have to wait for the population before triggering the webhooks.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Dedicated namespaces per test case

* Address PR review feedback

* Address PR review feedback

Co-authored-by: Tim Ebert <timebertt@gmail.com>
Co-authored-by: Tim Ebert <tim.ebert@sap.com>