Ensure that shoot node network and seed service network do not overlap.
The SNI implementation in Gardener using apiserver-proxy 'leaks' the cluster
ip of the kube-apiserver in the seed into all shoot nodes. As the ip is locally
assigned it can easily cause issues if the shoot node network overlaps with the
seed service network.
This change adds the corresponding validation and adapts the tests.