Fix `blackbox-exporter` probe for connectivity from seed
Co-Authored-By: Wesley Bermbach <wesley.bermbach@sap.com>

Before this commit, the `blackbox-exporter` sidecar container part of the `prometheus` pod running in the shoot namespace in the seed cluster was not able to resolve the `*.local.gardener.cloud` DNS names. Consequently, it was always failing the probe for the `kube-apiserver` connectivity in the seed cluster.
By including it into the `provider-local`'s `dnsconfig` webhook, the `prometheus` pod's is now pointed to the `coredns` running in the `gardener-extension-provider-local-coredns` namespace. This will make it able to resolve the `*.local.gardener.cloud` DNS names.