Overwrite migrate and restore operations for worker
Overwrites the Migrate and Restore operations for the worker resource to
call Delete and Reconcile. This is necessary because in the local setup
the shoot's worker node is created as a POD inside the shoot's
control plane namespace. During control plane migration it will be
deleted, because the control plane namespace on the source seed has to be
deleted. Then it will be recreated on the destination seed. Trying to
support a proper restoration in this case brings challenges and could be
addressed later on.