Fix `blackbox-exporter` checks in local setup (#6051)
* Fix `blackbox-exporter` probe for connectivity from shoot

Co-Authored-By: Istvan Ballok <istvan.zoltan.ballok@sap.com>
Co-Authored-By: Wesley Bermbach <wesley.bermbach@sap.com>

Before this commit, the `coredns` was only rewriting the `QUESTION SECTION` of DNS requests but not the `ANSWER SECTION`. For example:

```
$ dig api.local.local.internal.gardener.cloud

[...]

;; QUESTION SECTION:
;api.local.local.internal.gardener.cloud.			IN	A

;; ANSWER SECTION:
istio-ingressgateway.istio-ingress.svc.cluster.local.		283	IN	A	10.2.184.123

[...]
```

However, some applications (like envoy) do not accept such behaviour and treat it as MITM attack. Consequently, the `envoy` container part of the `apiserver-proxy` `DaemonSet` running in the shoot cluster was not able to open connections to the seed cluster.
By using `answer auto` the `ANSWER SECTION` is rewritten as well which makes `envoy` accepting the requests. See also https://coredns.io/plugins/rewrite/#response-rewrites for more details.

The upgrade to `coredns@1.9.1` was required to use the `answer auto` feature, however it makes sense to upgrade in general (we already use `1.9` for "regular" shoot clusters).

* Fix `blackbox-exporter` probe for connectivity from seed

Co-Authored-By: Wesley Bermbach <wesley.bermbach@sap.com>

Before this commit, the `blackbox-exporter` sidecar container part of the `prometheus` pod running in the shoot namespace in the seed cluster was not able to resolve the `*.local.gardener.cloud` DNS names. Consequently, it was always failing the probe for the `kube-apiserver` connectivity in the seed cluster.
By including it into the `provider-local`'s `dnsconfig` webhook, the `prometheus` pod's is now pointed to the `coredns` running in the `gardener-extension-provider-local-coredns` namespace. This will make it able to resolve the `*.local.gardener.cloud` DNS names.