Add node local dns option to shoot spec and roll nodes in case it is switched.
Node local dns mitigates a lot of the dns issues in a kubernetes environment.
Therefore, it should become standard going forward. However, there may be
scenarios where it is not required/desired. This is why we want to be able to
turn it on/off as desired.
This change adds the flags related to node local dns to the shoot specification.
The annotations, which were previously used, still work to keep compatibility.
Node local dns is enabled if either the annotation or option in the shoot
specification is turned on. The options for enforcing tcp for cluster/upstream
communication work the other way round, i.e. if either the annotation or the
option in the shoot specification is turned off the result will be that tcp is
not enforced.
In addition to the shoot specification changes, the nodes are now also rolled
if node local dns is switched as it may leave node configuration behind. This
will only become real once the provider extensions use this change.
To be able to later keep node local dns pods running on the nodes until the
nodes are rolled this change furthermore introduces a label to the nodes
indicating whether node local dns is enabled or not for this node. It also
adds a node selector to the node local dns daemonset. This will later, once
the provider extensions adopt the change, keep the pods on the right nodes.
For now, it is not changing the status quo.