`ControlPlane` controller deploys `local-path-provisioner`
Similar to what kind does to provide PVCs, the `ControlPlane` controller of `provider-local` now deploys the `local-path-provisioner` (see https://github.com/rancher/local-path-provisioner) with the exact same manifests/configuration.
Note that the name of the `ServiceAccount` and `ConfigMap` are hard-coded in the `local-path-provisioner` code, so they cannot be renamed (I tried just naming them `local-path-provisioner` like all the other resources).