[GEP-18] Credentials rotation e2e test (#6072)
* Allow reusing existing shoot in e2e tests

Co-Authored-By: Rafael Franzke <rafael.franzke@sap.com>

* Restructure credentials rotation e2e test

Preparation for adding more assertions for other rotated credentials

Co-Authored-By: Rafael Franzke <rafael.franzke@sap.com>

* Enhance precision of CA rotation assertions

Add more assertions about CA secrets on seed cluster managed by gardenlet.
Similar to the assertions for provider-local secrets to be confident we don't
accidentally change any CAs where they should not change.

Co-Authored-By: Rafael Franzke <rafael.franzke@sap.com>

* Add label to etcd encryption config secret

Helps to find the secret in the e2e test

Co-Authored-By: Rafael Franzke <rafael.franzke@sap.com>

* Explicitly enable static token kubeconfig

We want to assert that the kubeconfig is rotated in the e2e test.
Hence, we need to explicitly enable it.

Co-Authored-By: Rafael Franzke <rafael.franzke@sap.com>

* Add assertions for other rotated credentials

Co-Authored-By: Rafael Franzke <rafael.franzke@sap.com>

* Please the linter

* Relax ETCDEncryptionKeyVerifier

Don't assert that there is only 1 encryption config, there might be more if it
was recently rotated. Take the most recent one instead of the first one in
alphabetical order.

* Fix wrong assertion in CAVerifier.AfterPrepared

* Fix decoding of etcd encryption config

* Fix CAVerifier.AfterCompleted

* Don't use HaveKeyWithValue on secretConfigNamesToSecrets

Failure messages are incredibly hard to understand and truncated because
they are too long. This will make things at least a little bit better.

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>