Use static shoot name for credentials rotation test
This is done to make the shoot API server accessible to the test just like in
the default test, where we also use the admin kubeconfig from the subresource
to talk to the shoot.
This is a prerequisite for verifying that the kubeconfig is valid in all phases
of the credentials rotation process.