[NodeLocalDNS]: Do not consider annotations when determining worker pool hash/node roll outs.
The alpha api of node local dns did not rotate nodes. This could leave them in a bad state
in case of turning the feature off. However, to prevent unintended node roll outs the alpha
api is not considers when calculting whether a node roll out is required or not. In
particular, this allows cluster with the annotation set for node local dns to continue to
work without an unnecessary node roll out.