[GEP-18] Advanced checks for credentials e2e test (1) (#6098)
* Use static shoot name for credentials rotation test

This is done to make the shoot API server accessible to the test just like in
the default test, where we also use the admin kubeconfig from the subresource
to talk to the shoot.
This is a prerequisite for verifying that the kubeconfig is valid in all phases
of the credentials rotation process.

* Verify kubeconfig has the same CA bundle as in .data[ca.crt]

* Verify ca-cluster secret has the same bundle as ca-bundle in seed

* Verify shoot access using static token and admin kubeconfig

Co-Authored-By: Rafael Franzke <rafael.franzke@sap.com>

* Verify secret encryption

Co-Authored-By: Rafael Franzke <rafael.franzke@sap.com>

* Verify shoot access using client certificate requested via CSR

Co-Authored-By: Rafael Franzke <rafael.franzke@sap.com>

* Only clean up for existing shoot

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>