Keep node local dns components during the roll out of new nodes.
DNS is crucial to kubernetes. Therefore, it is important that the node local
dns components stay in place on the (old) nodes until the node is being rolled.
With this change, the node local dns components stay in the cluster until no
node has the corresponding label requiring node local dns. In the following
reconcile the resources will be removed.
(The pods are only scheduled on the proper nodes using a node selector already.)