Fix the "Target" variable of the vpa-recommendations dashboard (#6129)
* Export the vpa-recommendations in the current schema version

Besides the schema version, there is no other change in this commit.

-  "schemaVersion": 18,
+  "schemaVersion": 27,

* Load the suggested values for the Target variable on dashboard load

Previously, the suggested values for the Target variable were only
refreshed when the dependent variable "targetKind" was changed and hence
the suggestion list was incorrect when the dashboard was opened.

With this change the Target variable is going to be refreshed when the
dashboard is opened.