Adds ability to create second seed in the local setup (#6059)
* Makes metrics bind address configurable

Makes the MetricsBindAddress option configurable for the local-provider
extension controler so that multiple local-provider extension controllers
can be started on the same host. The address is kept as :8080 by
defautl.

* Overwrite migrate and restore operations for worker

Overwrites the Migrate and Restore operations for the worker resource to
call Delete and Reconcile. This is necessary because in the local setup
the shoot's worker node is created as a POD inside the shoot's
control plane namespace. During control plane migration it will be
deleted, because the control plane namespace on the source seed has to be
deleted. Then it will be recreated on the destination seed. Trying to
support a proper restoration in this case brings challenges and could be
addressed later on.

* Adds ability to create a second local seed

* Adds documentation

* Enable CopyEtcdBackupsDuringControlPlaneMigration feature gate in the local setup

* Apply review comments to docs

* Apply review comments

* Adds dedicated backup secret and changes directory of kubeconfig for second seed

* Fix nil pointer exception during backupentry reconciliation

* Restructure kustomize folders