Improve the CPU and memory usage calculation on the Node Details dashboard (#6132)
* Reexport the node-details dashboard

Only the plugin version is changed.

* Change the way the memory and CPU usage on the node is calculated

Previously, the memory usage was calculated based on the
allocatable and active memory metrics.
Now it is calculated based on the total and available memory metrics.

It turns out that the (total - available) metric is a better indication
of memory pressure. When the available memory approaches 0, the system
will experience OOM.

The CPU usage was also adjusted to use the total capacity instead of the
allocatable capacity.

See #6031.

Co-authored-by: Daniel Foehr <daniel.foehr@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

Co-authored-by: Daniel Foehr <daniel.foehr@sap.com>