Write the blackbox exporter scrape probe logs also to stdout (#6140)
Previously, the blackbox exporter wrote logs to stdout on the info level.

The logs of the scrape probes, e.g. regarding the requests to check the
connectivity to the k8s api server, were not logged to stdout.
They were available for operators only via port-forwarding to the web UI
of the blackbox exporter on port 9115.

After this change, the probe logs are also written to stdout so that
when operators are looking for the reason of probe failures, they
can more conveniently check it with `k logs blackbox-exporter-...` instead
of port-forwarding to the component.

Note that the blackbox exporter UI has 2 minor advantages
when checking the logs:
- The blackbox exporter UI allows to check the logs and metrics of
  individual scrape probe runs separately. Operators have to rely on the
  timestamp of the logs in the stdout stream to keep logs from different
  scrape runs apart.
- The caller attribute of the logs is masked in stdout, main.go:211 is
  reported for each log line. (Not really relevant for our use case.)
  See https://github.com/prometheus/blackbox_exporter/blob/d26fcc665935e962ba9ce331d3dc82babe0627c4/main.go#L211