Adapt network policies to reversed vpn. (#6143)
With reversed vpn only the vpn-seed-server pod in the cluster control plane has layer 3 routing
access to the shoot cluster. The kube-apiserver uses the vpn-seed-server as http proxy. Monitoring
utilizes kube-apiserver as another level of indirection.
This is why this change removes the annotations from kube-apiserver and prometheus in case reversed
vpn is active.
As there is basically no longer connectivity to the shoot networks it is also no longer necessary
to exclude the shoot networks from the private network range (allow-to-private-networks). This
allows additional scenarios where private IPv4 addresses are used in the shoot cluster as well as
for other services.