Allow cilium to be used on seeds. (#6130)
Kubernetes network policies are interpreted slightly different by different CNIs.
While calico works well with IP based policies, cilium expects IP ranges to be
cluster external. This is why the network policy, which should have made it possible
for the SNI istio-ingress gateway to forward traffic to a kube-apiserver in the
shoot control plane, was not working as expected causing traffic to be dropped.
Adding the istio-ingress gateway explicitly works well with cilium.