Add a recording rule to store the new metrics under their old name
These metrics are used on the dashboards that are contributed by the
gardener extensions.

Due to updating the kube-state-metrics version these metrics are renamed
which would "break" those dashboards as long as the adjustment of the
extensions is not released and deployed.

The extensions follow an independent release cycle so this mitigation
helps to avoid breaking those dashboards by temporarily storing the
new metrics under their old names.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Ismail Alidzhikov <i.alidjikov@gmail.com>