Introduce seed-local container image registry
We will use this registry as a mirror for shoots such that the images
built by Skaffold are accessible from the shoot cluster in case it gets
registered as `ManagedSeed`.

This is needed because we don't want to push the images built by
Skaffold to any official, publicly available registry.

In the future, we might even be able to reuse this such that we can
speed up the image pull processing times.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>