[provider-local] Add support for `ManagedSeed`s including e2e tests (#6293)
* `garden` namespace deployment is only needed for second kind cluster

In the first kind cluster, the `garden` namespace already exists because it runs the Gardener control plane.
Without this, the second client-side apply removes the project labels from the `Namespace`.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Enable `ShootVPAEnabledByDefault` admission plugin in local setup

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Simplify gardenlet bootstrap kubeconfig

This makes it usable for locally created `ManagedSeed`s and follows the same pattern like in `example/gardener-local/gardenlet/values-kind2.yaml`

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Add documentation for `ManagedSeed`s

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Add example Kubernetes resources for local `ManagedSeed`s

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Register `node` webhook for shoot clusters

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Introduce seed-local container image registry

We will use this registry as a mirror for shoots such that the images
built by Skaffold are accessible from the shoot cluster in case it gets
registered as `ManagedSeed`.

This is needed because we don't want to push the images built by
Skaffold to any official, publicly available registry.

In the future, we might even be able to reuse this such that we can
speed up the image pull processing times.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Allow machine pods to talk to the registry

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Mutate `containerd` config to import additional configuration files

This only applies to newly created nodes.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Bump `machine-controller-manager-provider-local` image

This includes
https://github.com/gardener/machine-controller-manager-provider-local/commit/f2c93198c794afc4e8e742a26026584ecce9aadf
which allows machine pods to talk to the seed API server. In the local
setup, the seed API server is also the garden API server and the
gardenlet needs to talk to it to register the `Seed`.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Mount `backup-path` volume with `DirectoryOrCreate` mode in `provider-local`

This will create the directory if it does not exist which is the case for shoot clusters registered as `ManagedSeed`s.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Explicitly delete `Ingress`es in seed deletion

This effectively fixes https://github.com/gardener/gardener/issues/6062.

It's actually a work around since not all seed system components (like
the monitoring stack) are deployed via `ManagedResource`s yet. Hence,
`gardener-resource-manager` does not clean this up for us and we have to
delete the resources manually.
We only do it for `Ingress`es now to fix above mentioned bug since the
deployment of the monitoring stack is anyways planned to be refactored
with [GEP-19](https://github.com/gardener/gardener/pull/6151).

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Only delete seed system resources after all other resources are deleted

Otherwise, we might delete important system resources like `PriorityClass`es which can cause extensions to not come up anymore (e.g. after being scaled by VPA). This can result in a deadlock during seed deletion.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Make `defaultShoot` function reuseable in `e2e` test packages

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Add e2e test for `ManagedSeed`s

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Drop testmachinery-based `ManagedSeed` test

This is no longer valuable now that we have an e2e test which can run on each PR and periodically on `master` branch.

Co-Authored-By: Tim Ebert <tim.ebert@sap.com>

* Address PR review feedback

* Address PR review feedback

Co-authored-by: Tim Ebert <tim.ebert@sap.com>