Differentiate the vpa metrics for the seed and control planes
See #6268, #1863

Previously, the 2 recording rules had the same name:
  `seed:hvpa_status_applied_vpa_recommendation_target:sum`
and caused a conflict because the resulting
time series had the same name and labels.

Now they are differentiated by the name suffix.

There were 2 other recording rules with the same name:
  `record: seed:hvpa_status_applied_vpa_recommendation_target:sum`
and the control plane version is similarly renamed now.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>