Add dashboard for monitoring conntrack race failures.
Connection tracking is an important optimization of the linux kernel to reduce the overhead
of applying network filter rules to all network packets.
Unfortunately, the connection tracking code in the linux kernel has some race conditions,
which can lead to connection attempts loosing their connection tracking entry. This is
especially possible when network address translation is used. The result is a failed
connection attempt.
This is why this change introduces two dashboards for the corresponding metric.