Restrict the amount of work being done in a privileged container in the reversed vpn client. (#6352)
It is not necessary for the vpn-seed-server to run as privileged container. The tun
device required for vpn operation can be mounted into the container namespace. The
remaining work requiring privileged work can be moved into an init container.
In case the container should run as non-root user in the future, it might help to
create a separate tun device in a privileged init container as described in
https://community.openvpn.net/openvpn/wiki/UnprivilegedUser#TUNTAPDevice .