Add documentation and integration test for `CloudProfile` controller (#6360)
* Add documentation

* Move `AddToManager` to dedicated `add.go` file

* Add integration test

* Remove pointer for config

Specifying a nil config is not a valid thing to do.
Otherwise, `AddToManager` would panic.
Hence, rather take a zero value config and take care of correctly handling unset fields.
This is semantically better, as we can't be sure that the given config struct is a real
(i.e. defaulted) object, e.g. in integration tests we don't do defaulting for the config.

* Improve integration test

Co-authored-by: Tim Ebert <timebertt@gmail.com>