Improve the ApiServerUnreachableViaKubernetesService alert
The prometheus in the shoot control plane namespace running in the seed
scrapes the blackbox-exporter that runs in the shoot cluster's kube-system
namespace.

The blackbox exporter tries to connect to the kubernetes api server
that runs in the control plane via the cluster internal service:
`kubernetes.default.svc.cluster.local`.

See the blue path on the diagram here:
https://gardener.cloud/docs/gardener/monitoring/connectivity/

The prometheus instance in the control plane can not reach the
blackbox-exporter in the shoot cluster directly, it uses
the kubernetes api server as a proxy (which uses the VPN connection
between the seed and shoot clusters to talk to the pod in the shoot
cluster).

See https://github.com/gardener/gardener/blob/a4506ca6ae52b9b312814f2449f28e2f5cde8c42/charts/seed-monitoring/charts/core/charts/prometheus/templates/config.yaml#L371-L404

The probe_success metric is 1 if the scrape was successful, i.e. if the
proxied connection to the blackbox-exporter and the connection from the
blackbox-exporter back to the kubernetes api server all work.

The probe_success metric is 0 if the proxied connection to the
blackbox-exporter works, but the blackbox-exporter can not reach the
kubernetes api server. This could happen if there is a network issue
in the shoot cluster but actually the kubernetes api server is working.

If the kubernetes api server is not working at all, than the proxied
connection will not work either and the scrape itself fails, meaning that
the probe_success series will be "absent" and become eventually stale.

For testing purposes, this error condition can be triggered by scaling
the kubernetes api server deployment in the control plane to 0.

  k scale deployment kube-apiserver --replicas=0

We'd expect that the ApiServerUnreachableViaKubernetesService alert
would fire in this case, but it doesn't. The promql expression
  `probe_success{job="blackbox-exporter-k8s-service-check"} != 1`
explicitly filters for a value that does not equal to 1,
and in this case the time series does not exist so the alert
will not fire.

One approach to improve this would be to use
  `absent(probe_success{job="blackbox-exporter-k8s-service-check"} == 1)`
but in that case the job and instance labels of the alert would be
missing, because the absent() function carries over the labels of the
instant-vector selector only if there is no function call or binary
operator involved.

See https://prometheus.io/docs/prometheus/latest/querying/functions/#absent

The slightly more complicated expression in this commit checks for either
explicit failure (== 0) or for the absence of the time series.
This way the job and instance labels (that need to be duplicated
unfortunately) are carried over by absent() to the alert, so
that the alert carries these labels in both failure cases.

Regarding the unit tests, please refer to the documentation first:
https://prometheus.io/docs/prometheus/latest/configuration/unit_testing_rules/#unit-testing-for-rules

The existing test with '0+0x30' checked the case that the probe_success
series had 0 values.

The newly added test case '1 1 _x10 stale' checks the new scenario when
the kubernetes api server crashes:
the probe success has a value 1 in the beginning (i.e. the connectivity
check is working),
then the scrape fails and eventually (after 5 minutes) the time series
becomes stale.
We expect an ApiServerUnreachableViaKubernetesService alert in this case.

Regarding staleness, see
https://prometheus.io/docs/prometheus/latest/querying/basics/#staleness
https://promcon.io/2017-munich/slides/staleness-in-prometheus-2-0.pdf

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>