Remove readiness/liveness probes and metrics requests from apiserver-proxy access log. (#6553)
The readiness/liveness probes in conjunction with the regular metrics requests spam
the access log of the apiserver-proxy to the point that it is hard to find the actual
workload requests. This change simply removes the readiness/liveness probes and
metrics requests from the access log so that the actual access log can focus on the
forwarded workload requests.