E2E auto-rotation test
First, we update the kubeconfig validity config and make gardenlet issue a new certificate. The values are chosen such that auto-rotation should happen every ~1m.
After a successful test, we revert the validity settings again (they will then be implicitly picked up with the next rotation) to not intervene with the `ManagedSeed` deletion and following assertions (we don't want too many gardenlet restarts).