Validate seed network configuration on gardenlet startup.
In case the seed specification does not fit to the actual seed cluster configuration
the result may be hard to find issues, e.g. in the network policies. Therefore, this
change introduces a validation during the gardenlet startup to compare the seed
specification with the reality in the seed cluster.
For now, this only works in case the seed cluster is managed itself by Gardener as
it uses the shoot-info config map. Other clusters are always accepted.