Prevent confusing error messages about outdated cloud configs on nodes during rolling updates (#6555)
* Cosmetics

* Map pools to entire secret metadata instead of checksum only

This is a prerequisite for a subsequent commit in which we will use more data from the secret's metadata to improve the accuracy of the `CloudConfigUpdatedForAllWorkerPools` function.

* Drop backwards-compatibility/migration code

* `nodeToBeDeleted` function can return an error

In a subsequent commit, this function will be able to return an error, so let's adapt it's signature accordingly already now so that the next commit is cleaner and easier to review.

* Skip nodes whose OSC key is different from their secret's OSC key

When the OSC key of a worker pool changes (e.g. when the Kubernetes version is upgraded) then a rolling update of the nodes is triggered.

Earlier, existing nodes (waiting for their removal (which can take some time)) were still considered by this health check function we're modifying here. Hence, the health check failed and complained that the cloud config is outdated. However, this is actually the desired behaviour since old nodes/to-be-deleted nodes are not supposed to update their cloud config.

With this change, we will consider nodes whose OSC key does not match the OSC key of their cloud config secret as 'to-be-deleted soon'. Consequently, the health check is skipped for them, preventing the misleading error messages.