Add option to automatically rewrite some dns requests to reduce amount of requests being made due to dns search path and ndots=5. (#6192)
* Add option to automatically rewrite some dns requests to reduce amount of requests being made due to dns search path and ndots=5.

Per default kubernetes uses a fairly long search path to make it easy for
developers/applications to use different short names for addressing services.
In conjunction with the fairly high value for ndots=5, which specifies how
many dots have to appear in a name for it to be seen as fully qualified,
this leads to a lot of dns requests being generated for a simple name like
kubernetes.default.svc.cluster.local. Depending on the infrastructure it might
be 10 requests (5 search path entries * 2 IP versions).
This is wasteful and may even lead to running into rate limits on the
infrastructure when the incorrect requests are forwarded the infrastructure
dns server. It is especially bad as for some requests it is fairly easy to
derive the "real" name that should be resolved, e.g.
kubernetes.default.svc.cluster.local.default.svc.cluster.local should
obviously resolve kubernetes.default.svc.cluster.local.
This pull request adds a gardenlet feature gate to automatically rewrite some
of the dns requests, which are obviously incorrect, to the most likely desired
ones. In addition to automatic rewrites for duplicates of the cluster domain, e.g.
cluster.local, it also provides means to rewrite some common domain suffixes.
This can help prevent search path iteration for commonly used domains, but
may have undesired effects if there is overlap to other requests. This is
why the common domain suffixes should contain at least two dots.
Please note that it is also possible to omit the search path at all if a
dot was to be added to the domain name. However, this is hardly ever done
in the real world/applications. Hence, this change might mitigate/improve
a lot of real world scenarios.
This includes documentation about the trade-offs and alternatives with
regards to dns search path optimizations.
An admission plugin is included to configure common suffixes automatically
depending the landscape.

* Addressed review feedback

* Added annotation to disable core dns query rewriting per shoot

* Fix potential import cycle/import restrictions due to usage of feature gates in helper function.