Fix to-dns network policy in garden namespace for node-local-dns if non-ip based policy engine is used.
Network policies are enforced by different means depending on the CNI. Some rely
on IPs, others work mostly on pod identities and ignore IPs mostly. (Calico falls
into the first category. Cilium is an example for the second category.)
To allow Seeds to run in both environments even with node-local-dns being enabled
the to-dns network policy must allow connections to the pod and its IP. The former
was missing here.