Update envoy proxy to v1.23.1 (#6366)
* Update envoy proxy to v1.23.0.

The new envoy proxy release is a bit more strict concerning the typing of its
configuration. Therefore, the config types have to be stated now explicitly.
As the envoy proxy distroless container image supports arm now, the container
image was also switched to reduce the attack surface. However, it is running
as non-root user, which required subsequent changes as well for vpn-seed-server
and apiserver-proxy.
As the switch to a non-root user showed that the capability NET_BIND_SERVICE
was not required by vpn-seed-server, all capabilities were removed from the
envoy proxy in vpn-seed-server.

* Update envoy proxy to v1.23.1