Extend `ManagedSeed` e2e test with gardenlet kubeconfig rotation scenarios (#6568)
* E2E rotation test via annotating `ManagedSeed`

* E2E rotation test via annotating kubeconfig `Secret`

* Enhance gardenlet config API with kubeconfig validity settings

* Earlier, it was not possible to configure dedicated validity. The global `--cluster-signing-duration` value of KCM was used. This only works for garden clusters >= 1.22 since it relies in the `.spec.expirationSeconds` field in the `CertificateSigningRequest` API only available in these versions.
* Earlier, the boundaries for the auto rotation jitter values were hard-coded to 70/90 and not configureable.

* Respect new gardenlet kubeconfig validity config API fields

* E2E auto-rotation test

First, we update the kubeconfig validity config and make gardenlet issue a new certificate. The values are chosen such that auto-rotation should happen every ~1m.
After a successful test, we revert the validity settings again (they will then be implicitly picked up with the next rotation) to not intervene with the `ManagedSeed` deletion and following assertions (we don't want too many gardenlet restarts).

* Use static shoot name to enable access via Prow

* Address PR review feedback