Add zone information to seed namespace
After etcd is ready we assume that at least one pod with a volume has been scheduled. This is the earliest point in time when we can "pin" the zone, i.e. all remaining volumes also reside in that zone and for the future it needs to be assured that all pods (esp. the ones with volumes) are scheduled there.