[gardener-local] Add pull-through cache registry (#6591)
* Switch to registry.k8s.io

With this, we reduce the number of upstream registries we use.

Co-Authored-By: Rafael Franzke <rafael.franzke@sap.com>

* Deploy pull through caches and use it in local seeds and shoots

The local registry is now also exposed in the KIND_ENV=local setup for consistency reasons.
It is not used by the setup currently, though.

Co-Authored-By: Rafael Franzke <rafael.franzke@sap.com>

* Parallelize image pulls in local clusters

Co-Authored-By: Rafael Franzke <rafael.franzke@sap.com>

* Enhance docs

* Allow configuring `KubeletConfiguration.Registry{PullQPS,Burst}`

If the params are given in the shoot spec, use them.
Otherwise, don't set them in the config file and let kubelet default them.

* Increase image pull rate in local clusters

* Add validation for new fields

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>