Improve the Node/Worker Pool Overview dashboard (#6595)
The query of the panel
"Average # of Pods on each Node in Worker Group(s) $worker_group"
was incorrect: it was off by 1.

The new query uses count instead of sum because actually we want to
count the pods.

When the 2 vectors are merged with the + operator to enrich the series
with the worker group label, we should add 0 instead of 1 to avoid
manipulating the result.

Note that we used the new Prometheus feature to format this query:

```
avg by (label_worker_gardener_cloud_pool) (
    count by (node) (kube_pod_info{type="shoot"})
  + on (node) group_right ()
    sum by (label_worker_gardener_cloud_pool, node) (
      0 * kube_node_labels{label_worker_gardener_cloud_pool=~"etcd.+"}
    )
)
```

The query of the panel
"Number of Nodes in Worker Group(s) $worker_group"
could be simplified.

The query of the panel
"Average Memory Usage of Nodes in Worker Group(s) $worker_group"
was slightly improved.

The query of the panel
"Average CPU Usage of Nodes in Worker Group(s) $worker_group"
was improved to exclude the iowait and steal CPU times.

The "Failed Connection Tracking Inserts" panel was missing the
rate function for the counter metric.

The queries of the "Overview of all nodes" table were also formatted
and improved in a similar way.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>