Refactor a prometheus relabeling rule in an equivalent way
The previous rule dropped metrics that had an empty container label.
The current rule keeps the metrics that have a non-empty container label.

The 2 rules are equivalent.

Motivation for this change: the golang regular expressions do not support negative
lookaheads.
So if we want to drop everything where the container label is empty, but
not all of those, we can not express that with a regular expression.
On the other hand, if we keep everything where the container label is
not empty, we can extend that regular expression with new patterns.

See https://prometheus.io/docs/prometheus/latest/configuration/configuration/#relabel_config

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>