Reduce cardinality of network I/O metrics
Enhance containerd network I/O metrics.
Add a table next to the network I/O panel that shows whether a pod is in
the host network namespace. This is important because if pods are in the
host network namespace, we cannot collect meaningful metrics. The host
network includes traffic from system services and other pods.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>