Keep network metrics only for the eth0 interface of pods
If a pod is not using the host network namespace, its main network
interface is called eth0. We should keep metrics of this eth0 interface.

If a pod is using the host network namespace, then all the network
interfaces of the host (ens.+, tunl.+, cali.+, cilium, ...) are going
to be reported for the pod. However, this is not so meaningful, because
these metrics do not capture network traffic that is specific to this pod,
but rather all the traffic in the host network namespace, which might come
from other pods and system services.

So with this change we only keep the network metrics for the pods
that are not in the host network namespace, and only for the eth0
network device.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>